Conducted refactoring in AprgCommon module.

+++ AprgCommon/AprgCommon/src/File/AlbaFileReader.cpp
@@ -3,10 +3,10 @@
 #include <Bit/AlbaBitManipulation.hpp>
 #include <String/AlbaStringHelper.hpp>
 
+#include <cstdint>
 #include <fstream>
 #include <numeric>
 #include <string>
-
 using namespace std;
 
 namespace alba
@@ -94,29 +94,35 @@ NumberType AlbaFileReader::getData()
     NumberType result(0);
     m_stream.read(getCharacterBufferPointer(), numberOfBytesToRead);
     unsigned int numberOfCharacters = static_cast<unsigned int>(m_stream.gcount());
-    result = accumulate(m_characterBuffer.cbegin(), m_characterBuffer.cbegin()+static_cast<int>(numberOfCharacters), static_cast<NumberType>(0U), [&](NumberType partialSum, NumberType newValue)
+    result = accumulate(m_characterBuffer.cbegin(), m_characterBuffer.cbegin()+static_cast<int>(numberOfCharacters),
+                        static_cast<NumberType>(0U),
+                        [&](NumberType partialSum, NumberType newValue)
     {
         partialSum = static_cast<NumberType>(partialSum << 8);
-        partialSum |= (0xFF & newValue);
-        return partialSum;
+        partialSum |= (0xFF & newValue);        return partialSum;
     });
     return result;
 }
 
-template unsigned int AlbaFileReader::getTwoByteData<unsigned int>();
-template unsigned int AlbaFileReader::getFourByteData<unsigned int>();
-template unsigned long long AlbaFileReader::getEightByteData<unsigned long long>();
-template unsigned int AlbaFileReader::getTwoByteSwappedData<unsigned int>();
-template unsigned int AlbaFileReader::getFourByteSwappedData<unsigned int>();
-template unsigned long long AlbaFileReader::getEightByteSwappedData<unsigned long long>();
-template unsigned int AlbaFileReader::getData<unsigned int, 2>();
-template unsigned int AlbaFileReader::getData<unsigned int, 4>();
-template unsigned int AlbaFileReader::getData<unsigned int, 8>();
+template uint16_t AlbaFileReader::getTwoByteData<uint16_t>();
+
+template uint32_t AlbaFileReader::getFourByteData<uint32_t>();
+
+template uint64_t AlbaFileReader::getEightByteData<uint64_t>();
+
+template uint16_t AlbaFileReader::getTwoByteSwappedData<uint16_t>();
+
+template uint32_t AlbaFileReader::getFourByteSwappedData<uint32_t>();
+
+template uint64_t AlbaFileReader::getEightByteSwappedData<uint64_t>();
+
+template uint16_t AlbaFileReader::getData<uint16_t, 2>();
+template uint32_t AlbaFileReader::getData<uint32_t, 4>();
+template uint64_t AlbaFileReader::getData<uint64_t, 8>();
 
 void AlbaFileReader::saveDataToMemoryBuffer(AlbaMemoryBuffer& buffer, unsigned int numberOfBytesToRead)
 {
-    char* writer = static_cast<char*>(buffer.resizeWithAdditionalSizeAndReturnBeginOfAdditionalData(numberOfBytesToRead));
-    m_stream.read(writer, static_cast<streamsize>(numberOfBytesToRead));
+    char* writer = static_cast<char*>(buffer.resizeWithAdditionalSizeAndReturnBeginOfAdditionalData(numberOfBytesToRead));    m_stream.read(writer, static_cast<streamsize>(numberOfBytesToRead));
 }
 
 string AlbaFileReader::getLineAndIgnoreWhiteSpaces()
